refactor: move styling to .styled file
